moved functionality from sidebar to page render
